refactor(domain): Fare, AgeFarePolicy, DistanceFarePolicy는 노선 및 구간에 대한 요금 도메인이므로, 패키지를 line으로 리팩토링(피드백 반영)
