+++ src/main/java/nextstep/subway/line/domain/AgeFarePolicy.java
@@ -1,4 +1,4 @@
-package nextstep.subway.fare.domain;
+package nextstep.subway.line.domain;
 
 import java.util.Arrays;
 import nextstep.subway.member.domain.Age;
