+++ src/test/java/nextstep/subway/line/domain/FareTest.java
@@ -1,11 +1,13 @@
-package nextstep.subway.fare.domain;
+package nextstep.subway.line.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import nextstep.subway.common.constant.ErrorCode;
+import nextstep.subway.line.domain.AgeFarePolicy;
 import nextstep.subway.line.domain.Distance;
+import nextstep.subway.line.domain.Fare;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
