+++ src/test/java/nextstep/subway/path/domain/PathTest.java
@@ -7,8 +7,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.Arrays;
 import java.util.Collections;
 import nextstep.subway.common.constant.ErrorCode;
-import nextstep.subway.fare.domain.AgeFarePolicy;
-import nextstep.subway.fare.domain.Fare;
+import nextstep.subway.line.domain.AgeFarePolicy;
+import nextstep.subway.line.domain.Fare;
 import nextstep.subway.line.domain.Distance;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
